first cut refactoring dpk_text_encoder
