+++ transforms/language/text_encoder/Dockerfile.python
@@ -20,20 +20,9 @@ RUN  pip install data-processing-dist/${DPK_WHEEL_FILE_NAME}
 
 # END OF STEPS destined for a data-prep-kit base image 
 
-COPY --chown=dpk:root pyproject.toml pyproject.toml
+COPY --chown=dpk:root dpk_text_encoder/ dpk_text_encoder/
 COPY --chown=dpk:root requirements.txt requirements.txt
-RUN pip install ${PIP_INSTALL_EXTRA_ARGS} --no-cache-dir -e .
-
-# copy transform main() entry point to the image 
-COPY ./src/text_encoder_transform_python.py .
-COPY ./src/text_encoder_transform.py .
-
-# copy some of the samples in
-COPY ./src/text_encoder_local.py local/
-
-# copy test
-COPY test/ test/
-COPY test-data/ test-data/
+RUN pip install ${PIP_INSTALL_EXTRA_ARGS} --no-cache-dir -r requirements.txt
 
 # Set environment
 ENV PYTHONPATH /home/dpk
