+++ transforms/language/text_encoder/README.md
@@ -1,14 +1,140 @@
-# Text encoder Transform 
+# Text Encoder Transform 
 
-This transform is using sentence encoder models to create embeddings of the text in each row of the input .parquet table.
+Please see the set of
+[transform project conventions](../../../README.md#transform-project-conventions)
+for details on general project conventions, transform configuration,
+testing and IDE set up.
 
-Per the set of 
-[transform project conventions](../../README.md#transform-project-conventions)
-the following runtimes are available:
+## Contributors
 
-* [python](python/README.md) - provides the base python-based transformation 
-implementation.
-* [ray](ray/README.md) - enables the running of the base python transformation
-in a Ray runtime
-* [kfp](kfp_ray/README.md) - enables running the ray docker image 
-in a kubernetes cluster using a generated `yaml` file.
+- Michele Dolfi (dol@zurich.ibm.com)
+
+## Description 
+
+This transform is using [sentence encoder models](https://en.wikipedia.org/wiki/Sentence_embedding) to create embedding vectors of the text in each row of the input .parquet table.
+
+The embeddings vectors generated by the transform are useful for tasks like sentence similarity, features extraction, etc which are also at the core of retrieval-augmented generation (RAG) applications.
+
+### Input 
+
+| input column name | data type | description |
+|-|-|-|
+| the one specified in _content_column_name_ configuration | string | the content used in this transform |
+
+
+### Output columns
+
+
+| output column name | data type | description |
+|-|-|-|
+| the one specified in _output_embeddings_column_name_ configuration | `array[float]` | the embeddings vectors of the content |
+
+
+## Configuration
+
+The transform can be tuned with the following parameters.
+
+
+| Parameter  | Default  | Description  |
+|------------|----------|--------------|
+| `model_name`                    | `BAAI/bge-small-en-v1.5` | The HF model to use for encoding the text. |
+| `content_column_name`           | `contents` | Name of the column containing the text to be encoded. |
+| `output_embeddings_column_name` | `embeddings` | Column name to store the embeddings in the output table. |
+
+
+## Usage
+
+### Launched Command Line Options 
+
+When invoking the CLI, the parameters must be set as `--text_encoder_<name>`, e.g. `--text_encoder_column_name_key=myoutput`.
+
+
+### Running the samples
+To run the samples, use the following `make` targets
+
+* `run-cli-sample` - runs src/text_encoder_transform.py using command line args
+* `run-local-sample` - runs src/text_encoder_local.py
+
+These targets will activate the virtual environment and set up any configuration needed.
+Use the `-n` option of `make` to see the detail of what is done to run the sample.
+
+For example, 
+```shell
+make run-cli-sample
+...
+```
+Then 
+```shell
+ls output
+```
+To see results of the transform.
+
+### Code example
+
+TBD (link to the notebook will be provided)
+
+
+### Transforming data using the transform image
+
+To use the transform image to transform your data, please refer to the 
+[running images quickstart](../../../../doc/quick-start/run-transform-image.md),
+substituting the name of this transform image and runtime as appropriate.
+
+## Testing
+
+Following [the testing strategy of data-processing-lib](../../../../data-processing-lib/doc/transform-testing.md)
+
+Currently we have:
+- [Unit test](test/test_text_encoder_python.py)
+
+
+
+# TextEncoder Ray Transform 
+Please see the set of
+[transform project conventions](../../../README.md#transform-project-conventions)
+for details on general project conventions, transform configuration,
+testing and IDE set up.
+
+## Summary 
+This project wraps the [text_encoder transform](../python) with a Ray runtime.
+
+## Configuration and command line Options
+
+Text Encoder configuration and command line options are the same as for the base python transform. 
+
+## Running
+
+### Launched Command Line Options 
+In addition to those available to the transform as defined in [here](../python/README.md),
+the set of 
+[ray launcher](../../../../data-processing-lib/doc/ray-launcher-options.md) are available.
+
+### Running the samples
+To run the samples, use the following `make` targets
+
+* `run-cli-sample` - runs src/text_encoder_transform.py using command line args
+* `run-local-sample` - runs src/text_encoder_local_ray.py
+* `run-s3-sample` - runs src/text_encoder_s3_ray.py
+    * Requires prior installation of minio, depending on your platform (e.g., from [here](https://min.io/docs/minio/macos/index.html)
+     and [here](https://min.io/docs/minio/linux/index.html) 
+     and invocation of `make minio-start` to load data into local minio for S3 access.
+
+These targets will activate the virtual environment and set up any configuration needed.
+Use the `-n` option of `make` to see the detail of what is done to run the sample.
+
+For example, 
+```shell
+make run-cli-sample
+...
+```
+Then 
+```shell
+ls output
+```
+To see results of the transform.
+
+### Transforming data using the transform image
+
+To use the transform image to transform your data, please refer to the 
+[running images quickstart](../../../../doc/quick-start/run-transform-image.md),
+substituting the name of this transform image and runtime as appropriate.
