+++ transforms/language/text_encoder/dpk_text_encoder/ray/transform.py
@@ -24,7 +24,7 @@ from data_processing_ray.runtime.ray import RayTransformLauncher
 from data_processing_ray.runtime.ray.runtime_configuration import (
     RayTransformRuntimeConfiguration,
 )
-from text_encoder_transform import TextEncoderTransformConfiguration
+from dpk_text_encoder.transform import TextEncoderTransformConfiguration
 
 
 logger = get_logger(__name__)
