+++ transforms/language/text_encoder/kfp_ray/text_encoder_wf.py
@@ -20,7 +20,7 @@ from workflow_support.compile_utils import ONE_HOUR_SEC, ONE_WEEK_SEC, Component
 task_image = "quay.io/dataprep1/data-prep-kit/text_encoder-ray:latest"
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "text_encoder_transform_ray.py"
+EXEC_SCRIPT_NAME: str = "-m dpk_text_encoder.ray.transform"
 
 # components
 base_kfp_image = "quay.io/dataprep1/data-prep-kit/kfp-data-processing:latest"
