--- transforms/language/text_encoder/ray/.dockerignore
@@ -1,1 +1,0 @@
-venv/
