--- transforms/language/text_encoder/ray/test-data/expected/metadata.json
@@ -1,45 +1,0 @@
-{
-    "pipeline": "pipeline_id",
-    "job details": {
-        "job category": "preprocessing",
-        "job name": "text_encoder",
-        "job type": "pure python",
-        "job id": "job_id",
-        "start_time": "2024-08-05 11:07:20",
-        "end_time": "2024-08-05 11:07:23",
-        "status": "success"
-    },
-    "code": {
-        "github": "github",
-        "commit_hash": "12345",
-        "path": "path"
-    },
-    "job_input_params": {
-        "content_column_name": "contents",
-        "output_embeddings_column_name": "embeddings",
-        "model_name": "BAAI/bge-small-en-v1.5",
-        "checkpointing": false,
-        "max_files": -1,
-        "random_samples": -1,
-        "files_to_use": [".parquet"]
-    },
-    "job_output_stats": {
-        "source_files": 1,
-        "source_size": 1058,
-        "result_files": 1,
-        "result_size": 5389,
-        "processing_time": 0.38291406631469727,
-        "nfiles": 1,
-        "nrows": 2,
-        "source_doc_count": 2,
-        "result_doc_count": 2
-    },
-    "source": {
-        "name": "data-prep-kit/transforms/language/text_encoder/python/test-data/input",
-        "type": "path"
-    },
-    "target": {
-        "name": "data-prep-kit/transforms/language/text_encoder/python/output",
-        "type": "path"
-    }
-}
